test(lua,ruby): isolate RFC 8785 and RFC 3339 test suites into dedicated files
Per PR #102 review (inline comment 3298060956). Lua and Ruby previously
mixed canonical-JSON (RFC 8785) and expires-parser (RFC 3339) cases in
core_spec.lua / core_test.rb. Split into:

  lua/tests/json_canonical_rfc8785_spec.lua
  lua/tests/expires_rfc3339_spec.lua
  ruby/test/json_canonical_rfc8785_test.rb
  ruby/test/expires_rfc3339_test.rb

PHP already kept JsonTest.php and ChallengeTest.php separate. No change
needed on PHP. Lua run.lua updated to require both new spec files.
