+++ lua/tests/run.lua
@@ -8,6 +8,8 @@ package.path = table.concat({
 
 require('tests.network_check_spec')
 require('tests.core_spec')
+require('tests.json_canonical_rfc8785_spec')
+require('tests.expires_rfc3339_spec')
 require('tests.server_spec')
 require('tests.solana_verify_spec')
 require('tests.html_spec')
